rename *-service.ts to *.service.ts
To match the current structure (logger.service.ts for instance)
