+++ src/lib/logger.barrel.ts
@@ -3,24 +3,24 @@ export * from './config/iconfig-engine';
 export * from './config/config-engine';
 
 export * from './mapper/ilog-position';
-export * from './mapper/imapper-service';
-export * from './mapper/mapper-service';
+export * from './mapper/imapper.service';
+export * from './mapper/mapper.service';
 
-export * from './metadata/imetadata-service';
+export * from './metadata/imetadata.service';
 export * from './metadata/imetadata';
-export * from './metadata/metadata-service';
+export * from './metadata/metadata.service';
 
 export * from './monitor/ilogger-monitor';
 export * from './monitor/logger-monitor';
 
-export * from './rules/irules-service';
-export * from './rules/rules-service';
+export * from './rules/irules.service';
+export * from './rules/rules.service';
 
-export * from './server/iserver-service';
-export * from './server/server-service';
+export * from './server/iserver.service';
+export * from './server/server.service';
 
-export * from './writer/iwriter-service';
-export * from './writer/writer-service';
+export * from './writer/iwriter.service';
+export * from './writer/writer.service';
 export * from './writer/color-scheme';
 
 export * from './types/logger-level.enum';
