+++ src/lib/rules/rules.service.ts
@@ -2,7 +2,7 @@ import { DatePipe } from '@angular/common';
 import { Injectable } from '@angular/core';
 import { NgxLoggerLevel } from '../types/logger-level.enum';
 import { INGXLoggerConfig } from '../config/iconfig';
-import { INGXLoggerRulesService } from './irules-service';
+import { INGXLoggerRulesService } from './irules.service';
 
 @Injectable()
 export class NGXLoggerRulesService implements INGXLoggerRulesService {
